deploy/scripts/update_s3/requirements.txt
boto3
python-dotenv
mypy
GitPython
